Allows resetting filtered changes to their original values (where possible)
z_offset and input_shaper values will be automatically reset and the SAVE_CONFIG will continue with the reset values.

pid values and mesh values cannot be reset automatically and filtering them will result in the SAVE_CONFIG being canceled.

In the case where a SAVE_CONFIG is canceled and a SHUTDOWN is scheduled:
The shutdown will be delayed, giving the user the opportunity to stop it from shutting down and losing the pending changes